requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
async-timeout==4.0.3
attrs==23.1.0
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
diffusers==0.23.1
filelock==3.13.1
Flask==3.0.0
frozenlist==1.4.0
fsspec==2023.10.0
gunicorn==21.2.0
huggingface-hub==0.19.4
idna==3.6
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
lightning-utilities==0.10.0
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.4
networkx==3.2.1
numpy==1.26.2
omegaconf==2.3.0
packaging==23.2
Pillow==10.1.0
pytorch-lightning==2.1.2
PyYAML==6.0.1
regex==2023.10.3
request-id-flask==0.0.2
requests==2.31.0
safetensors==0.4.1
sympy==1.12
tokenizers==0.15.0
torch==2.1.1
torchmetrics==1.2.0
tqdm==4.66.1
transformers==4.35.2
typing_extensions==4.8.0
urllib3==2.1.0
Werkzeug==3.0.1
yarl==1.9.3
zipp==3.17.0
